Show property names if the generated name does not contain a p element
Because property names may be derived from the @var tag if no short description
is present is it possible that markup from the markdown generation may end up
in the short description. This causes paragraph elements to be displayed in
the summary.

To combat the paragraph elements I had added a special piece of code that
filters those. Unfortunately this meant that if there were no paragraph
elements that the entire string is filtered.

To combat this side effect I have added a check that the filtering only
happens if a paragraph element is actually encountered.