On error filter: hide results with no entries
In issue https://github.com/phpDocumentor/phpDocumentor2/issues/423 @GaryJones
mentioned that the report filter could be improved by hiding all files that
no longer had any errors associated with it.

This change introduces this behaviour and updates the error counter next to
the filename so that that accurately reflects the number of shown issues.