Visibility filters do not work properly
In issue https://github.com/phpDocumentor/phpDocumentor2/issues/573 @ferjul17 mentioned
that the visibility filters do not function properly.

During research we discovered that the logic was faulty and that bootstrap sets the 'active'
class after this event is called. With this commit we fix this behaviour and it works as
expected.